<rdar://problem/10466170> Regression: Scroll bars disappear and don't come back
Reviewed by Sam Weinig.

Fixing a merge issue. On TOT we use #if USE(SCROLLBAR_PAINTER), but on the branch
it is still #if USE(WK_SCROLLBAR_PAINTER). Also, on the branch, all of the
m_scrollbarPainterController functions go through WebKitSystemInterface instead of
using forward declaration. We should consider merging those changes, but in the
meantime, for this to work, we have to go back to the branch-style.
* platform/mac/ScrollAnimatorMac.mm:
(WebCore::ScrollAnimatorMac::notityPositionChanged):



Canonical link: https://commits.webkit.org/81538.165@safari-534.52-branch
git-svn-id: https://svn.webkit.org/repository/webkit/branches/safari-534.52-branch@100809 268f45cc-cd09-0410-ab3c-d52691b4dbfc