Merge r124259.
    2012-07-31  Anders Carlsson  <andersca@apple.com>

Prefer the Oracle Java plug-in over the Apple Java plug-in
https://bugs.webkit.org/show_bug.cgi?id=92780

Reviewed by Oliver Hunt.

* UIProcess/Plugins/mac/PluginInfoStoreMac.mm:
(WebKit::findPluginWithBundleIdentifier):
Add a new helper for finding a plug-in with the given bundle identifier.

(WebKit::checkForPreferredPlugin):
Helper function for making sure that an old plug-in is never loaded if a new plug-in is found, and that the old plug-in
is removed from the list of loaded plug-ins if the new plug-in is found.

(WebKit::PluginInfoStore::shouldUsePlugin):
Prefer the Oracle Java plug-in over the Apple Java plug-in.



Canonical link: https://commits.webkit.org/81538.476@safari-534.58-branch
git-svn-id: https://svn.webkit.org/repository/webkit/branches/safari-534.58-branch@130058 268f45cc-cd09-0410-ab3c-d52691b4dbfc