[macOS] Remove mdnsresponder access from WebKit processes https://bugs.webkit.org/show_bug.cgi?id=218633 <rdar://problem/69133548>
Reviewed by Per Arne Vollan.

Now that all network loading (including media) happens in the Network process,
we can remove an old permission to open outbound connections to mdnsresponder.
This is also unneeded in the GPU process, for the same reason.

I noticed that the new WebAuthn process sandbox mistakenly included this,
and removed it there, too.

* GPUProcess/mac/com.apple.WebKit.GPUProcess.sb.in:
* WebAuthnProcess/mac/com.apple.WebKit.WebAuthnProcess.sb.in:
* WebProcess/com.apple.WebProcess.sb.in:


Canonical link: https://commits.webkit.org/231295@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@269491 268f45cc-cd09-0410-ab3c-d52691b4dbfc