Look at parents when event bubbles for input element activation behavior https://bugs.webkit.org/show_bug.cgi?id=218660
Reviewed by Ryosuke Niwa.

LayoutTests/imported/w3c:

Rebaseline WPT test that is now fully passing.

* web-platform-tests/dom/events/Event-dispatch-click-expected.txt:

Source/WebCore:

When a click event is dispatched at a node that is not an HTMLInputElement, and if the
event bubbles, we should look up the tree to see if there is an HTMLInputElement that
we should trigger activation behavior for. Prevously, we were failing to do this.

The new behavior is consistent with Blink.

No new tests, rebaselined existing test.

* dom/EventDispatcher.cpp:
(WebCore::findInputElementInEventPath):
(WebCore::EventDispatcher::dispatchEvent):


Canonical link: https://commits.webkit.org/231373@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@269582 268f45cc-cd09-0410-ab3c-d52691b4dbfc