Add microbenchmark for FinalizationRegistry basic usage https://bugs.webkit.org/show_bug.cgi?id=218667
Reviewed by Yusuke Suzuki.

The no unregister token benchmark runs in ~40ms and the unregister
token in ~25ms on my Mac Pro. Interestingly, it also shows that
using an unregister token make FinalizationRegistry ~4x slower
(total time is longer because the number of objects is 10x bigger
in the test). This is because we have to store all the data in a
HashMap<UnregisterToken, JSObject> rather than a Vector<JSObject>.

* microbenchmarks/finalizationRegistryCleanup.js: Added.
(async callback):
* microbenchmarks/finalizationRegistryCleanupWithUnregisterToken.js: Added.
(async callback):


Canonical link: https://commits.webkit.org/231381@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@269590 268f45cc-cd09-0410-ab3c-d52691b4dbfc