Stop using objc_autoreleasePool{Push, Pop} in RunLoop observers https://bugs.webkit.org/show_bug.cgi?id=218728
Reviewed by Tim Horton.

Use in a RunLoop observer is deprecated when a built-in version of the
behavior is available (because the conflicting pool management commands
crash each other).

This is UIWebView, so there are no rules. Tested manually, at a safe
distance.

* platform/ios/wak/WebCoreThread.mm:
(WebRunLoopLockInternal):
(WebRunLoopUnlockInternal):
(WebRunLoopEnableNested):
(WebRunLoopDisableNested):
(RunWebThread):


Canonical link: https://commits.webkit.org/231441@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@269658 268f45cc-cd09-0410-ab3c-d52691b4dbfc